git-extra: fix notepad usage for commit messages
Uh oh. It seems that the patch that got merged into `master` somehow
was not the patch that was tested, but instead a faulty one.

The problem was that `git config <key>` returns an error if the
key was not found and it made the code fail when no column wrapping
was configured (and besides, the test for the column wrapping was
inverted).

To prove that *this* commit is good, it was made using

	git -c core.editor=notepad -c format.commitmessagecolumns=72
	commit -s

This fixes https://github.com/git-for-windows/git/issues/517

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>