msi: include git-bash and git-cmd
Files that need to be copied into the top-level install directory need
to be handled slightly differently than the others, as we already found
out when we started including post-install.bat.

They need the KeyPath property, and they must not specify a relative
folder after "INSTALLFOLDER".

In contrast to the post-install.bat entry, we skip the Guid property
(which is used to state the the post-install.bat script is expected to
go away after the installation, which is not the case for git-cmd.exe
nor git-bash.exe).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>