Add a script to trigger GitLab's CI for specific remote refs
Since GitLab's Continuous Integration needs to be configured via an
in-tree configuration file, we need to add that locally and then push the
branch to GitLab's Git for Windows space to trigger a build.

The configuration is a *little* bit tricky, for a couple of reasons:

- we need to make sure that the `git.exe` built in a previous run is
  not used by the runner to update the repository, and

- we want to run the tests in parallel, ideally on many runners.

To address the first issue, we simply build and test in the ci/
subdirectory and *never* build a git.exe in the top-level directory.

To address the second wish, we use a custom "add-on" Makefile to
generate an archive with the artifacts required for testing. This
is required because we want to reuse the generated artifacts on multiple
runners. We also need to adjust the absolute paths in some, for that
reason.

To support this use case, we wrap the repetitive steps in a script which
we "commit" together with the .gitlab-ci.yml file that is required to
trigger a build on GitLab.

We also use the output of script that was committed in the parent commit
whose purpose is to partition the test scripts into a number of buckets
with roughly equal runtime.

Note that GitLab's runner on Windows does not really kill off the build,
but only the top-level process (on Windows, killing a process does not
kill that process' children...). Therefore, we use GNU timeout (which is
part of Git for Windows' SDK's coreutils package), with the KILL signal,
to make sure that the build is *really* terminated.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>