Add a script to partition Git's test into roughly equal chunks
This is a small node.js script taking the output of `prove --timer` as
stdin, and an optional number of buckets as command-line parameter.

It then partitions Git's tests into buckets of consecutive ranges such
that each range takes roughly the same time to run. We choose to go with
ranges to be resilient against newly-introduced tests: we do
not want to forget to run those!

The intended recipient is of course the GitLab CI runner which builds
Git for Windows on one node, and then runs the test suite, in parallel,
on as many nodes as we can get.

We use a Bresenham-like method to partition the tests, and sort the
buckets by descending runtime, to avoid being stuck waiting for a single
runner to finish off the slowest bucket at the end.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>