Add a script to commit the files of a Git for Windows SDK
To facilitate spawning build agents, let's have an easy-to-fetch (and
easy-to-cache) repository containing all the files of a Git for Windows
SDK.

The concept is similar to the olden days of msysGit (no, no, no, not Git
for Windows, msysGit was never Git for Windows, it was always the
development environment for the contributors who were willing to
develop Git for Windows). In those days, we abused Git as a package
management system.

The difference is that we have a proper package management system now,
and we simply synchronize the files into a Git repository so that build
agents have an easier time to keep up-to-date with the Git for Windows
SDK (because there is no chance that build agents can update some MSYS2
installation interactively, working around problems with bash, pacman or
the msys2-runtime being updated while in use).

This script is not in its final form: currently, it is only able to
perform an initial commit of all packages. We will have to teach it to
detect which packages were updated, and to remove the old files.

And then we will also have to teach the script which files are part of a
package even if they were not included verbatim in the .tar.xz archive
but generated by the .INSTALL script.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>