commit-msys2: move the current functionality into the 'init' command
We will want to keep the repository up-to-date with the local state
after running `pacman -Syu`, i.e. after updating the packages to their
respective newest versions. This will require a different functionality
entirely.

So let's introduce an 'init' command and move the current code to
initialize the repository into that command. While at it, make it a bit
more complete by ensuring that there *is* a repository, and by
committing the locally-cached package index, too.

The functionality to keep it up-to-date will be introduced later, as the
'commit' command.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>