git-extra: add a helper to edit shortcut files (WIP)
This helper is a close copy of create-shortcut.exe, with additional
options to modify the history buffer and the quick edit flags.

The intention was to help with the issue raised here (where an interactive
Python interpreter would not have a history available to itself):
https://github.com/git-for-windows/git/issues/1013

In my hands, this was a bit flakey when editing existing settings in the
shortcut file (sometimes it would pick up changes only after calling the
helper a couple of times).

The bigger problem, though, is that the defaults may very well be bogus.
For example, who knows what the default width and height is? Or for that
matter, the history size? We cannot just edit the history buffer size,
after all, but have to write an *entire* NT_CONSOLE_PROPS block.

Maybe this is not worth all the time that I spent on this issue after all.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>