Start implementing a meta credential helper
With this helper, configuring `credenial.helper = chooser` will let the
user choose which credential helper (of those found in the `PATH`, via
matching the shell pattern `git-credential-*`) to use.

It will also lets the user persist their choice into the system config,
making it an ideal default for the portable Git.

This commit only implements a "Choose" and a "Cancel" button, leaving
the reset to the next commits.

This will address https://github.com/git-for-windows/git/issues/2116

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>