installer: force-overwrite any values in the system config, if necessary
The Git Credential Manager Core installer adds entries to the system
config (an empty one to re-start the list of credential helpers, and
then the entry for `git-credential-manager-core.exe`).

This is fine, but when we try to install Git for Windows, it leads to an
error:

	Unable to set system config
	"credential.helper":="manager-core": exit code 5

	stderr:
	warning: credential.helper has multiple values
	error: cannot overwrite multiple values with a single value
	    Use a regexp, --add or --replace-all to change
	    credential.helper.

Since the system config is owned by Git for Windows, simply use
`--replace-all` when setting system config values.

Reported-by: Derrick Stolee <dstolee@microsoft.com>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>