Add an automated PR build
This build will automatically determine which package need to be built,
build them in a build matrix, and then publish the packages as build
artifacts.

Note: since we do not have any packages in build-extra which depend on
one another, we do not need to take precautions to build them in a
specific order and to then install the just-built dependencies.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>