PR build (run the installer): do verify the Git version
This is one of those copy/edit fails for which I am so righteously
famous: There simply _is no `bundle-artifacts/` directory in this
workflow and therefore the code currently greps for the empty string,
which is trivially found in the `version` file.

Instead, record the intended Git version in the step that creates the
sdk-artifact, and then use that information to validate that the
intended Git version was installed.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>